requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
rich>=13.0.0
